requirements.txt
log4mongo==1.7.0
requests==2.20.0
pycryptodome==3.12.0